Changes to Kaprekar:
On the first run through the program I have encountered 3 Errors within kaprekarUtils.c:
Line 29: in the test case ensuring that the program will automatically return false when a negative output is examined, the program literally attempts to return 'false' rather than return by a declared test case. This was edited to return 0.
Error Number 2 was in regards to line 39 and the declared variable "numDigits." It is fairly clear that the wrote "numberOfDigits" instead of his declared variable "numDigits."
The third error falls in regard to the variable "modulus." on line 43. The variable was incorrectly spelled "modulous," causing the final red flag.